feat: refactor attendance handling to use centralized attendance action and geolocation utility
