+++ frontend/src/community/attendance/components/molecules/AttendanceModals/ClockOutModal/ClockOutModal.tsx
@@ -2,21 +2,16 @@ import { Box, Stack, Typography, useTheme } from "@mui/material";
 import { ButtonV2 } from "@rootcodelabs/skapp-ui";
 import { FC } from "react";
 
-import { useUpdateEmployeeStatus } from "~community/attendance/api/AttendanceApi";
-import { useGetAttendanceConfiguration } from "~community/attendance/api/AttendanceAdminApi";
 import { useAttendanceStore } from "~community/attendance/store/attendanceStore";
 import { AttendanceSlotType } from "~community/attendance/types/attendanceTypes";
 import {
   calculateWorkedDuration,
   calculateWorkedDurationInHoursAndMinutes
 } from "~community/attendance/utils/CalculateWorkedDuration";
+import { useAttendanceAction } from "~community/attendance/hooks/useAttendanceAction";
 import Icon from "~community/common/components/atoms/Icon/Icon";
-import { appModes } from "~community/common/constants/configs";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { IconName } from "~community/common/types/IconTypes";
-import { convertDateToUTC } from "~community/common/utils/dateTimeUtils";
-import { useUpdateEmployeeStatusWithLocation } from "~enterprise/attendance/api/AttendanceApi";
-import { useGetEnvironment } from "~enterprise/common/hooks/useGetEnvironment";
 
 import styles from "./styles";
 
@@ -39,42 +34,11 @@ const ClockOutModal: FC<Props> = ({ closeModal }) => {
 
   const translateText = useTranslator("attendanceModule", "timeWidget");
 
-  const { isPending, mutate } = useUpdateEmployeeStatus();
-  const { mutate: mutateWithLocation, isPending: isEpPending } =
-    useUpdateEmployeeStatusWithLocation();
-
-  const environment = useGetEnvironment();
-  const isEnterprise = environment === appModes.ENTERPRISE;
-
-  const { data: attendanceConfig } = useGetAttendanceConfiguration();
-  const isGeoFencingEnabled = attendanceConfig?.isGeoFencingEnabled ?? false;
+  const { recordAttendance, isPending } = useAttendanceAction();
 
   const handleProceedHome = (): void => {
     const slotType = setSlotType(AttendanceSlotType.END);
-
-    if (isGeoFencingEnabled && isEnterprise) {
-      navigator.geolocation.getCurrentPosition(
-        (position) => {
-          mutateWithLocation({
-            recordActionType: slotType,
-            time: convertDateToUTC(new Date().toISOString()) as string,
-            latitude: position.coords.latitude,
-            longitude: position.coords.longitude
-          });
-        },
-        () => {
-          mutateWithLocation({
-            recordActionType: slotType,
-            time: convertDateToUTC(new Date().toISOString()) as string,
-            latitude: null,
-            longitude: null
-          });
-        }
-      );
-    } else {
-      mutate(slotType);
-    }
-
+    recordAttendance(slotType);
     closeModal();
   };
 
@@ -131,7 +95,7 @@ const ClockOutModal: FC<Props> = ({ closeModal }) => {
             <ButtonV2
               onClick={handleProceedHome}
               aria-label={translateText(["confirm"])}
-              isLoading={isPending || isEpPending}
+              isLoading={isPending}
               variant={"primary"}
               icon={<Icon name={IconName.RIGHT_ARROW_ICON} />}
               iconPosition="end"
